Fixed some bugs + added Colors for Dragon names.

+++ src/main/java/de/tobiyas/enderdragonsplus/damagewhisperer/DragonDamageWhisperer.java
@@ -14,7 +14,7 @@ public class DragonDamageWhisperer implements Runnable{
 	private static LinkedList<EnderDragon> damageList = new LinkedList<EnderDragon>();
 	
 	private EnderDragon dragon;
-	private int dragonHealth;
+	private double dragonHealth;
 	private String damager;
 	
 	private int tickerTime = 5;
@@ -38,22 +38,22 @@ public class DragonDamageWhisperer implements Runnable{
 		Player player = Bukkit.getPlayer(damager);
 		if(player == null)
 			return;
-		int newHealth = dragon.getHealth();
+		double newHealth = dragon.getHealth();
 		if(newHealth < dragonHealth)
 			broadcastDamage(player, dragonHealth - newHealth);
 	}
 	
-	private void broadcastDamage(Player player, int damage){
-		int actualLife = dragon.getHealth();
-		int maxLife = dragon.getMaxHealth();
+	private void broadcastDamage(Player player, double damage){
+		double actualLife = dragon.getHealth();
+		double maxLife = dragon.getMaxHealth();
 		
 		String midLifeString = parsePersentageLife(actualLife, maxLife);
 		player.sendMessage(ChatColor.YELLOW + "The Dragon has " + midLifeString + ChatColor.YELLOW + " health left. You did " + 
 				ChatColor.LIGHT_PURPLE + damage + ChatColor.YELLOW + " damage.");
 	}
 	
-	private String parsePersentageLife(int actual, int max){
-		float currentPercentage = actual / max;
+	private String parsePersentageLife(double actual, double max){
+		double currentPercentage = actual / max;
 		
 		if(currentPercentage < 0.2)
 			return ChatColor.RED + "" + actual + "/" + max;
